adds s to the end of the newly renamed getAllDataGroupAction
