+++ src/login/components/GroupManagement/GroupManagementContainer.js
@@ -1,7 +1,7 @@
 import { connect } from "react-redux";
 import GroupManagement from "./GroupManagement";
 import * as createGroupActions from "../../redux/actions/createGroupActions";
-import * as allDataActions from "../../redux/actions/getAllGroupMgmtDataAction";
+import * as allDataActions from "../../redux/actions/getAllGroupMgmtDataActions";
 import i18n from "../../translation/InjectIntl_HOC_factory";
 
 const mapStateToProps = (state, props) => {
